+++ src/etcd3aio/__init__.py
@@ -14,13 +14,14 @@ from .errors import (
 )
 from .kv import SortOrder, SortTarget, prefix_range_end
 from .lease import LeaseKeepalive
-from .maintenance import AlarmType
+from .maintenance import AlarmType, DowngradeAction
 from .watch import WatchFilter
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 __all__ = [
     'AlarmType',
+    'DowngradeAction',
     'Election',
     'Etcd3Client',
     'EtcdConnectionError',